# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WhatsApp Gateway SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send single message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [sent, failed, pending]
        deviceId:
          type: string
        deviceName:
          type: string
        mediaUrl:
          type: string
        createdAt:
          type: string
      required: [id, phone, message, status, deviceId, createdAt]
    SendMessageRequest:
      type: object
      properties:
        phone:
          type: string
        message:
          type: string
        deviceId:
          type: string
        mediaUrl:
          type: string
      required: [phone, message, deviceId]
...